Add:新增地圖

+++ ptsd-template-main/src/Menu.cpp
@@ -16,8 +16,15 @@ void Menu::Move(glm::vec2 displacement, glm::vec2 goal) {
 bool Menu::IfFocus() {
     int Cursor_X = Util::Input::GetCursorPosition()[0];
     int Cursor_Y = Util::Input::GetCursorPosition()[1];
-    int Menu_X_Range[2] = {(this->GetPosition()[0] - (this->GetScaledSize()[0] / 2)), (this->GetPosition()[0] + (this->GetScaledSize()[0] / 2))};
-    int Menu_Y_Range[2] = {(this->GetPosition()[1] - (this->GetScaledSize()[1] / 2)), (this->GetPosition()[1] + (this->GetScaledSize()[1] / 2))};
+    int Menu_X_Range[2] = {
+        static_cast<int>(this->GetPosition()[0] - (this->GetScaledSize()[0] / 2.0f)),
+        static_cast<int>(this->GetPosition()[0] + (this->GetScaledSize()[0] / 2.0f))
+    };
+    
+    int Menu_Y_Range[2] = {
+        static_cast<int>(this->GetPosition()[1] - (this->GetScaledSize()[1] / 2.0f)),
+        static_cast<int>(this->GetPosition()[1] + (this->GetScaledSize()[1] / 2.0f))
+    };
     if((Cursor_X > Menu_X_Range[0]) && (Cursor_X < Menu_X_Range[1])) {
         if ((Cursor_Y > Menu_Y_Range[0]) && (Cursor_Y < Menu_Y_Range[1])) {
             return true && GetVisibility();
